fix(stella-cli): finish #4775's split — main has not compiled since it landed
`main` does not build. #4775 split `command_deck.rs` into sibling modules by
COPYING items into the siblings and adding the imports, without deleting the
originals. Twenty-three names ended up defined twice, and the first of them
stops the compiler:

    error[E0255]: the name `handle_supervisor_msg` is defined multiple times

The canary filed #4810 against `ce6b1cf3c` (#4799's merge), but that commit
is not the cause: its diff to `command_deck.rs` is three lines of its own,
and `dccd5eb65` — the commit before it — carries the same duplicates. A walk
back through main puts the first broken tree at `5bad8ddd7`, #4775 itself.
Every PR since has been red for this reason, and the hold check now blocks
every merge, which is why this is labelled `unblocks-main`.

The compiler reported one E0255 and masked the rest; a scan of every
top-level item in `command_deck.rs` against every sibling found 23.

**The fix is not "delete the duplicates".** Four of the parent's copies are
NEWER than the siblings', because PRs based on a pre-#4775 main kept editing
them and their merges kept both:

- `spawn_mcp_connect` — the parent has #4798's `McpPlan::Servers(servers,
  notices)` plugin notices; the sibling still has the single-field arm.
- `run_deck_command` — the parent routes through `command_side::run` (the
  queue-free commands); the sibling still has the old inline `/help` arm.
- `handle_agents_input` / `tool_policy_inbound` — path and rustfmt drift.

Deleting the parent's copies would have silently reverted two shipped
features. So each item's live body is ported into the sibling that owns it,
then removed from the parent; where the two were identical the parent's copy
is simply dropped. `command_deck.rs` goes 3735 -> 2536 lines, which is what
#4775 was for.

Three follow-on repairs the port surfaced, none of them weakening anything:

- `command_side` and `authoring` join `slash_commands`'s module imports, and
  `ModelsCommand` / `parse_models_command` become `pub(super)` so
  `command_side` can still reach them.
- `turn_files`'s three driver ledgers named `command_deck.rs` as the file
  owning `run_lead_turn`. It has been `command_deck/lead_turn.rs` since
  #4775, so the rows are repointed — the guard still asserts the driver's own
  file names the seam, and now asserts it about the right file. These three
  would have gone red the moment anything compiled.
- `crates/stella-tui/src/v2/session.rs` carried unformatted code already on
  main; `cargo fmt --all` picks it up here so `format-check` can pass.

Verified on this tree: `cargo check -p stella-cli --bins` exits 0,
`cargo clippy -p stella-cli --all-targets -- -D warnings` exits 0,
`cargo test -p stella-cli --bins` is 2286 passed / 0 failed, rustdoc is clean
in both forms, and `make guards-fast` exits 0.

Closes #4810
