+++ crates/stella-cli/src/command_deck/slash_commands.rs
@@ -11,11 +11,11 @@ use stella_tools::ToolRegistry;
 use stella_tui::{Inbound, WorkspaceInput};
 use tokio::sync::mpsc::UnboundedSender;
 
-use super::authoring::{
-    agents_list_inbound, delete_agent, pin_agent, record_agent_invocation, save_agent,
-};
+use super::authoring::agents_list_inbound;
 use super::panel_snapshots::engine_config_inbound;
-use super::{LEAD, add_dir, init_cmd, model_cmd, profile_cmd, settings_io, skills, theme_cmd};
+use super::{
+    LEAD, add_dir, authoring, command_side, init_cmd, model_cmd, profile_cmd, settings_io, skills,
+};
 use crate::config::Config;
 use crate::interactive::AskUserIo;
 
@@ -51,7 +51,7 @@ pub(super) enum DeckCommand {
 /// a single recognized token (plus `refresh --force`); anything
 /// sentence-like stays a prompt, matching the "`/init do the thing` is a
 /// model prompt" rule.
-enum ModelsCommand {
+pub(super) enum ModelsCommand {
     /// `/info refresh [--force]` — re-sync the catalog, no model call.
     Refresh { force: bool },
     /// `/info list` — the same listing the bare `/info` prints.
@@ -63,7 +63,7 @@ enum ModelsCommand {
 
 /// Parse `trimmed` as a [`ModelsCommand`]; `None` leaves it on the normal
 /// path (custom expansion, then prompt).
-fn parse_models_command(trimmed: &str) -> Option<ModelsCommand> {
+pub(super) fn parse_models_command(trimmed: &str) -> Option<ModelsCommand> {
     let (head, rest) = trimmed.split_once(char::is_whitespace)?;
     let rest = rest.trim();
     if !matches!(head, "/info" | "/models") || rest.is_empty() {
@@ -103,7 +103,7 @@ pub(super) fn handle_agents_input(
             scope,
             content,
         } => {
-            let status = save_agent(root, name, *scope, content);
+            let status = authoring::save_agent(root, name, *scope, content);
             let _ = in_tx.send(agents_list_inbound(root, Some(status)));
             true
         }
@@ -112,12 +112,12 @@ pub(super) fn handle_agents_input(
             scope,
             version,
         } => {
-            let status = pin_agent(root, name, *scope, *version);
+            let status = authoring::pin_agent(root, name, *scope, *version);
             let _ = in_tx.send(agents_list_inbound(root, Some(status)));
             true
         }
         WorkspaceInput::AgentDelete { name, scope } => {
-            let status = delete_agent(root, name, *scope);
+            let status = authoring::delete_agent(root, name, *scope);
             let _ = in_tx.send(agents_list_inbound(root, Some(status)));
             true
         }
@@ -131,7 +131,7 @@ pub(super) fn handle_agents_input(
 ///
 /// Vocabulary: `/help`, `/clear`, `/info`, `/model`, `/init`, `/agents`.
 /// `/files`, `/diff`, `/graph` are deck-local (tab switches) and
-/// consumed TUI-side; an unknown bare `/command` gets a hint rather than a
+/// consumed in interactive mode; an unknown bare `/command` gets a hint rather than a
 /// wasted model call. Every productized command is no-argument, so the
 /// *whole* trimmed input is matched — `/init do the thing` is a model prompt,
 /// not a silent reindex that discards the rest. Custom commands/skills (⚡)
@@ -165,13 +165,14 @@ pub(super) async fn run_deck_command(
             event: AgentEvent::Text { text },
         });
     };
+    // The queue-free commands live in `command_side`, shared with the two
+    // `WorkspaceInput::Command` arms so a mid-turn `/export` and a queued
+    // one from an old journal run the same code. Asked first: whatever it
+    // recognizes never reaches the arms below.
+    if command_side::run(trimmed, cfg, in_tx, session_id) {
+        return DeckCommand::Handled;
+    }
     match trimmed {
-        "/help" => {
-            // Open the same rich, scrollable overlay the `?` key opens —
-            // every key, every tab, every slash command in one place. Far
-            // more useful (and readable) than a cramped one-line summary.
-            let _ = in_tx.send(Inbound::ShowHelp);
-        }
         "/clear" => {
             // Reset the driver's own LLM history…
             messages.clear();
@@ -186,17 +187,12 @@ pub(super) async fn run_deck_command(
         }
         // Bare `/model` is normally consumed deck-side (it opens the session
         // model picker); a queued or replayed one lands here and gets the
-        // textual summary instead of silence.
+        // textual summary instead of silence. NOT queue-free
+        // (`command_side`): `/model <spec>` below switches the running
+        // session, so the whole command stays on the turn-coupled path
+        // rather than having its bare form answer from a different place.
         "/model" => {
             say(model_cmd::current_summary(cfg));
-        }
-        // `/info` — provider/model information (`/models` before the rename;
-        // the old name still routes, undiscoverably, like `/tasks` → `/plan`).
-        "/info" | "/models" => {
-            say(Config::available_models_plain(None));
-        }
-        "/theme" => {
-            say(theme_cmd::current_summary(cfg));
         }
         "/init" => {
             // The splash replay, the narrator, and the question channel all
@@ -216,32 +212,10 @@ pub(super) async fn run_deck_command(
                 Err(e) => say(format!("init failed: {e}")),
             }
         }
-        // Export THIS session's telemetry to a timestamped ZIP archive of raw
-        // JSON dumps + a self-contained HTML dashboard. The session id is what
-        // scopes it: the archive is built to be shared, and #2558 records what
-        // shipping the whole workspace store cost.
-        "/export" => say(crate::export::export_command(&cfg.workspace_root, session_id).await),
         "/reload" => say(settings_io::reload_command(cfg, in_tx)),
-        "/donate" => {
-            say("❤️  Support Stella\n\
-                 \n\
-                 Stella is free, open-source, and local-first — no server, no \
-                 account, no telemetry sent home. If it's saving you time or \
-                 money, consider becoming a GitHub Sponsor:\n\
-                 \n\
-                   → https://github.com/sponsors/macanderson\n\
-                 \n\
-                 Recurring sponsorships keep development sustainable. You'll \
-                 see the available tiers and perks (one-time and monthly) on \
-                 that page. Every pledge helps fund the next feature, the next \
-                 provider, and the next release.\n\
-                 \n\
-                 Thank you! 🙏"
-                .to_string());
-        }
         // Deck-local commands (tab switches, `/agents` opening the Agents
-        // tab, the transcript-page overlays) are normally consumed TUI-side,
-        // but a queued one reaches here — accept it as handled (a no-op)
+        // tab, the transcript-page overlays) are normally consumed in
+        // interactive mode, but a queued one reaches here — accept it as handled (a no-op)
         // rather than calling it "unknown".
         "/files" | "/diff" | "/graph" | "/agents" | "/agent" | "/skills" | "/mcp"
         | "/mcp-search" | "/settings" | "/sessions" | "/subagents" | "/context" | "/inspect"
@@ -293,41 +267,6 @@ pub(super) async fn run_deck_command(
                 }
                 return DeckCommand::Handled;
             }
-            // `/theme <slug>` — switch + persist the colour theme (parity with
-            // `/model`). The live switch is a buffer remap in `stella_tui`, so
-            // it lands on the next frame; here we just flip it and save.
-            if let Some(command) = theme_cmd::parse_theme_command(trimmed) {
-                match command {
-                    theme_cmd::ThemeCommand::Set(name) => match theme_cmd::set_theme(name) {
-                        Ok(msg) | Err(msg) => say(msg),
-                    },
-                    theme_cmd::ThemeCommand::Usage(arg) => say(theme_cmd::usage(&arg)),
-                }
-                return DeckCommand::Handled;
-            }
-            // The `/info` argument forms first (see [`ModelsCommand`]):
-            // handled model-free — a catalog refresh is part of digging out
-            // of a broken model setting, so it can never be allowed to
-            // depend on a working model.
-            if let Some(command) = parse_models_command(trimmed) {
-                match command {
-                    ModelsCommand::Refresh { force } => {
-                        say("Model catalog refresh…".to_string());
-                        let mut emit = |line: String| say(line);
-                        if let Err(e) =
-                            crate::model_catalog::run_refresh_emit(force, &mut emit).await
-                        {
-                            say(format!("refresh failed: {e}"));
-                        }
-                    }
-                    ModelsCommand::List => say(Config::available_models_plain(None)),
-                    ModelsCommand::Usage(word) => say(format!(
-                        "`/info {word}` — unknown subcommand; try `/info` or `/info list` \
-                         (the listing) or `/info refresh [--force]` (re-sync the catalog)"
-                    )),
-                }
-                return DeckCommand::Handled;
-            }
             // A custom command/skill/agent (⚡): expand its template —
             // arguments and all — into the prompt the model turn runs.
             // Reserved names never reach a custom definition (`/init do the
@@ -335,7 +274,7 @@ pub(super) async fn run_deck_command(
             // An AGENT invocation additionally records a usage-telemetry
             // row (agent, pinned version, task) on the registry's ledger.
             if let Some(expanded) = custom.expand(trimmed, &skills::deck_reserved()) {
-                record_agent_invocation(trimmed, custom, registry);
+                authoring::record_agent_invocation(trimmed, custom, registry);
                 return DeckCommand::Expanded(expanded);
             }
             // A bare unknown /word is a typo'd command, not a prompt — say so
